+++ python-sdk/src/forta_agent/utils.py
@@ -1,20 +1,11 @@
-import base64
 import sys
 import os
 from jsonc_parser.parser import JsoncParser
 import requests
-import datetime
-import time
 from web3.auto import w3
 from web3 import Web3
-import json
-import logging
-
 from .forta_graphql import AlertsResponse
 
-DISPTACHER_ABI = [{"inputs":[{"internalType":"uint256","name":"agentId","type":"uint256"},{"internalType":"uint256","name":"scannerId","type":"uint256"}],"name":"areTheyLinked","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"view","type":"function"}]
-DISPATCH_CONTRACT = "0xd46832F3f8EA8bDEFe5316696c0364F01b31a573"; # Source: https://docs.forta.network/en/latest/smart-contracts/
-MOCK_JWT = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJib3QtaWQiOiIweDEzazM4N2IzNzc2OWNlMjQyMzZjNDAzZTc2ZmMzMGYwMWZhNzc0MTc2ZTE0MTZjODYxeWZlNmMwN2RmZWY3MWYiLCJleHAiOjE2NjAxMTk0NDMsImlhdCI6MTY2MDExOTQxMywianRpIjoicWtkNWNmYWQtMTg4NC0xMWVkLWE1YzktMDI0MjBhNjM5MzA4IiwibmJmIjoxNjYwMTE5MzgzLCJzdWIiOiIweDU1NmY4QkU0MmY3NmMwMUY5NjBmMzJDQjE5MzZEMmUwZTBFYjNGNEQifQ.9v5OiiYhDoEbhZ-abbiSXa5y-nQXa104YCN_2mK7SP0'
 
 def get_web3_provider():
     from . import web3Provider
@@ -105,7 +96,8 @@ def get_alerts(dict):
     query_options = AlertQueryOptions(dict)
     payload = query_options.get_query()
 
-    response = requests.request("POST", forta_api, json=payload, headers=headers)
+    response = requests.request(
+        "POST", forta_api, json=payload, headers=headers)
 
     if response.status_code == 200:
         data = response.json().get('data')
@@ -140,91 +132,3 @@ def hex_to_int(strVal):
 
 def keccak256(val):
     return Web3.keccak(text=val).hex()
-
-def fetch_jwt(claims={}, expiresAt=None) -> str:
-    host_name = 'forta-jwt-provider'
-    port = 8515
-    path = '/create'
-
-    uri = 'http://{host_name}:{port}{path}'.format(host_name=host_name, port=port, path=path)
-
-    if( (expiresAt != None) and (isinstance(expiresAt, datetime.datetime) == False)):
-        raise Exception("expireAt must be of type datetime")
-
-    if(expiresAt is not None):
-        exp_in_sec = expiresAt.timestamp()
-        claims["exp"] = exp_in_sec
-
-    if(os.environ['NODE_ENV'] != 'production'):
-        return MOCK_JWT
-
-    
-    response = requests.request("POST", uri, json={'claims': claims})
-
-    if response.status_code == 200:
-        data = response.json()
-        return data.get('token')
-
-    else:
-        raise Exception("Error occured with response fetching jwt token.", response)
-
-
-def verify_jwt(token: str, polygonUrl: str ='https://polygon-rpc.com') -> bool:
-    splitJwt = token.split('.')
-    rawHeader = splitJwt[0]
-    rawPayload = splitJwt[1]
-
-    header = json.loads(base64.urlsafe_b64decode(rawHeader + '==').decode('utf-8'))
-    payload = json.loads(base64.urlsafe_b64decode(rawPayload + '==').decode('utf-8'))
-
-    alg = header['alg']
-    botId = payload['bot-id']
-    expiresAt = payload['exp']
-    signerAddress = payload['sub']
-
-    if (signerAddress is None) or (botId is None):
-        logging.warning('Invalid claim')
-        return False
-    
-    if alg != 'ETH':
-        logging.warning('Unexpected signing method: {alg}'.format(alg=alg))
-        return False
-
-    currentUnixTime = time.mktime(datetime.datetime.utcnow().utctimetuple())
-
-    if expiresAt < currentUnixTime:
-        logging.warning('Jwt expired')
-        return False
-
-    msg = '{header}.{payload}'.format(header=rawHeader, payload=rawPayload)
-    msgHash = w3.keccak(text=msg)
-    b64signature = splitJwt[2]
-    signature = base64.urlsafe_b64decode(f'{b64signature}=').hex()
-    recoveredSignerAddress = w3.eth.account.recoverHash(msgHash, signature=signature)
-
-    if recoveredSignerAddress != signerAddress:
-        logging.warn('Signature invalid: expected={signerAddress}, got={recoveredSignerAddress}'.format(signerAddress=signerAddress, recoveredSignerAddress=recoveredSignerAddress))
-        return False
-
-    w3Client = Web3(Web3.HTTPProvider(polygonUrl))
-    contract = w3Client.eth.contract(address=DISPATCH_CONTRACT,abi=DISPTACHER_ABI)
-
-    areTheyLinked = contract.functions.areTheyLinked(int(botId,0), int(recoveredSignerAddress,0)).call()
-    
-    return areTheyLinked
-
-class DecodedJwt:
-    def __init__(self, dict):
-        self.header = dict.get('header')
-        self.payload = dict.get('payload')
-
-
-def decode_jwt(token):
-    # Adding need 4 byte for pythons b64decode
-    header = base64.urlsafe_b64decode(token.split('.')[0] + '==').decode('utf-8')
-    payload = base64.urlsafe_b64decode(token.split('.')[1] + '==').decode('utf-8')
-
-    return DecodedJwt({
-        "header": header,
-        "payload": payload
-    })
